upload hooks10 solution
took me about a week to do this, doubted my initial thought process which ended up being correct. simply look at all the hints and seen if the orthogonal values add up to the value, this can be done via combinations of the orthogonal values or simply summing them and comparing them against the hint value and seeing if it possible to achieve the sum. doing this along with all of the valid permutations of distributing the numbers throughout the hooks